Strict checking on file extension for site_files
Do not serve the .json (like fontspec.json and lexicon.json) and any other
file which doesn't look like an image or text or torrent.

+++ lib/AmuseWikiFarm/Controller/SiteFiles.pm
@@ -55,12 +55,18 @@ sub local_files :Chained('root') :PathPart('') :Args(1) {
     my ($self, $c, $file) = @_;
     # this method has already a sanity checker on the filename
     if (my $path = $c->stash->{site}->has_site_file($file)) {
-        $c->stash(serve_static_file => $path);
-        $c->detach($c->view('StaticFile'));
+        if ($file =~ m/[a-z0-9]
+                       \.
+                       (
+                           png | jpe?g | gif | ico | otf | ttf | woff |
+                           torrent | txt | css | js
+                       )\Z/x) {
+            $c->stash(serve_static_file => $path);
+            $c->detach($c->view('StaticFile'));
+            return;
+        }
     }
-    else {
-        $c->detach('/not_found');
-    }
+    $c->detach('/not_found');
 }
